预设系统全部下场,cmd系统独立,改进函数方块
