--- GTMBplugin_B/Presets/EULA.preset
@@ -1,34 +1,0 @@
-{
-    "autoCreate": false,
-    "bindParentAutoScale": false,
-    "bindParentOffset": [
-        0,
-        0,
-        0
-    ],
-    "boxDataList": [],
-    "childParts": [],
-    "childPresets": [],
-    "classType": "UIPreset",
-    "createUIMethod": "PushScreen",
-    "dimension": 0,
-    "forceLoad": true,
-    "isBindParent": false,
-    "isHud": true,
-    "mNode": null,
-    "name": "EULA",
-    "notifyRemove": true,
-    "objId": "0926b0eb057c4c80b2ea4d5c94641161",
-    "preLoad": true,
-    "presetId": "828ecb76-5632-4d72-94ac-d4ed24d06328",
-    "presetType": "界面预设",
-    "previewEnable": true,
-    "showInEditor": true,
-    "singleton": false,
-    "uiName": "EULA",
-    "uiNamespace": "ycgnw",
-    "uiNodeModule": "EULA",
-    "uiNodeModulePath": "gtmbPlugin\\uiScript\\EULA",
-    "uiNodeScreen": "GTMB_EULA.main",
-    "unique": false
-}
