--- GTMBplugin_B/Presets/cmdbatch.preset
@@ -1,34 +1,0 @@
-{
-    "autoCreate": false,
-    "bindParentAutoScale": false,
-    "bindParentOffset": [
-        0,
-        0,
-        0
-    ],
-    "boxDataList": [],
-    "childParts": [],
-    "childPresets": [],
-    "classType": "UIPreset",
-    "createUIMethod": "PushScreen",
-    "dimension": 0,
-    "forceLoad": true,
-    "isBindParent": false,
-    "isHud": true,
-    "mNode": null,
-    "name": "cmdbatch",
-    "notifyRemove": true,
-    "objId": "555cbc43ef094f5e886dd6f19742325f",
-    "preLoad": true,
-    "presetId": "4a666526-fb0b-485b-ab29-67a075f72c93",
-    "presetType": "界面预设",
-    "previewEnable": true,
-    "showInEditor": true,
-    "singleton": false,
-    "uiName": "cmdbatch",
-    "uiNamespace": "dtqyu",
-    "uiNodeModule": "cmdbatch",
-    "uiNodeModulePath": "gtmbPlugin\\uiScript\\cmdbatch",
-    "uiNodeScreen": "cmdbatch.main_closable",
-    "unique": false
-}
