--- GTMBplugin_B/Presets/enchant.preset
@@ -1,34 +1,0 @@
-{
-    "autoCreate": false,
-    "bindParentAutoScale": false,
-    "bindParentOffset": [
-        0,
-        0,
-        0
-    ],
-    "boxDataList": [],
-    "childParts": [],
-    "childPresets": [],
-    "classType": "UIPreset",
-    "createUIMethod": "PushScreen",
-    "dimension": 0,
-    "forceLoad": true,
-    "isBindParent": false,
-    "isHud": true,
-    "mNode": null,
-    "name": "enchant",
-    "notifyRemove": true,
-    "objId": "97307a8321d845bcab45a2d36e17a6f7",
-    "preLoad": true,
-    "presetId": "fce48a53-a388-4c19-9177-69ded741a577",
-    "presetType": "界面预设",
-    "previewEnable": true,
-    "showInEditor": true,
-    "singleton": false,
-    "uiName": "enchant",
-    "uiNamespace": "sblki",
-    "uiNodeModule": "enchantUI",
-    "uiNodeModulePath": "gtmbPlugin\\uiScript\\enchantUI",
-    "uiNodeScreen": "enchant.main_closable",
-    "unique": false
-}
