--- GTMBplugin_B/Presets/functionBlockScreen.preset
@@ -1,34 +1,0 @@
-{
-    "autoCreate": false,
-    "bindParentAutoScale": false,
-    "bindParentOffset": [
-        0,
-        0,
-        0
-    ],
-    "boxDataList": [],
-    "childParts": [],
-    "childPresets": [],
-    "classType": "UIPreset",
-    "createUIMethod": "PushScreen",
-    "dimension": 0,
-    "forceLoad": false,
-    "isBindParent": false,
-    "isHud": true,
-    "mNode": null,
-    "name": "functionBlockScreen",
-    "notifyRemove": true,
-    "objId": "8a090fa696cf429295ea6a50dd06a7ee",
-    "preLoad": false,
-    "presetId": "a63cd7e5-04aa-4201-a214-784af4823470",
-    "presetType": "界面预设",
-    "previewEnable": true,
-    "showInEditor": true,
-    "singleton": false,
-    "uiName": "functionBlockScreen",
-    "uiNamespace": "gtmb_plugin",
-    "uiNodeModule": "function_block_screen",
-    "uiNodeModulePath": "gtmbPlugin\\uiScript\\function_block_screen",
-    "uiNodeScreen": "function_block_screen.node_screen",
-    "unique": false
-}
