--- GTMBplugin_B/Presets/getitem.preset
@@ -1,34 +1,0 @@
-{
-    "autoCreate": false,
-    "bindParentAutoScale": false,
-    "bindParentOffset": [
-        0,
-        0,
-        0
-    ],
-    "boxDataList": [],
-    "childParts": [],
-    "childPresets": [],
-    "classType": "UIPreset",
-    "createUIMethod": "PushScreen",
-    "dimension": 0,
-    "forceLoad": true,
-    "isBindParent": false,
-    "isHud": true,
-    "mNode": null,
-    "name": "getitem",
-    "notifyRemove": true,
-    "objId": "c639ba9798414120a0fd622032fe7437",
-    "preLoad": true,
-    "presetId": "a6dfd5e9-4bc0-483b-a7b9-737c51dc89ed",
-    "presetType": "界面预设",
-    "previewEnable": true,
-    "showInEditor": true,
-    "singleton": false,
-    "uiName": "getitem",
-    "uiNamespace": "sblki",
-    "uiNodeModule": "getitemUI",
-    "uiNodeModulePath": "gtmbPlugin\\uiScript\\getitemUI",
-    "uiNodeScreen": "getitem.main_closable",
-    "unique": false
-}
