--- GTMBplugin_B/Presets/itemTips.preset
@@ -1,34 +1,0 @@
-{
-    "autoCreate": false,
-    "bindParentAutoScale": false,
-    "bindParentOffset": [
-        0,
-        0,
-        0
-    ],
-    "boxDataList": [],
-    "childParts": [],
-    "childPresets": [],
-    "classType": "UIPreset",
-    "createUIMethod": "PushScreen",
-    "dimension": 0,
-    "forceLoad": true,
-    "isBindParent": false,
-    "isHud": false,
-    "mNode": null,
-    "name": "itemTips",
-    "notifyRemove": true,
-    "objId": "03a0fc858ce94134afffb5556a078f8d",
-    "preLoad": true,
-    "presetId": "04043d9a-d4be-4f0c-80c6-73005a671f4e",
-    "presetType": "界面预设",
-    "previewEnable": true,
-    "showInEditor": false,
-    "singleton": false,
-    "uiName": "itemTips",
-    "uiNamespace": "ycgnw",
-    "uiNodeModule": "itemTips",
-    "uiNodeModulePath": "gtmbPlugin\\uiScript\\itemTips",
-    "uiNodeScreen": "customtips.main_closable",
-    "unique": false
-}
