--- GTMBplugin_B/Presets/listenBlockScreen.preset
@@ -1,34 +1,0 @@
-{
-    "autoCreate": false,
-    "bindParentAutoScale": false,
-    "bindParentOffset": [
-        0,
-        0,
-        0
-    ],
-    "boxDataList": [],
-    "childParts": [],
-    "childPresets": [],
-    "classType": "UIPreset",
-    "createUIMethod": "PushScreen",
-    "dimension": 0,
-    "forceLoad": false,
-    "isBindParent": false,
-    "isHud": true,
-    "mNode": null,
-    "name": "listenBlockScreen",
-    "notifyRemove": true,
-    "objId": "c54e599ea03840f7b31fb268ab6763f7",
-    "preLoad": false,
-    "presetId": "1341d964-f146-4800-8e39-45fae12af8af",
-    "presetType": "界面预设",
-    "previewEnable": true,
-    "showInEditor": true,
-    "singleton": false,
-    "uiName": "listenBlockScreen",
-    "uiNamespace": "gtmb_plugin",
-    "uiNodeModule": "listenBlockScreen",
-    "uiNodeModulePath": "gtmbPlugin\\uiScript\\listenBlockScreen",
-    "uiNodeScreen": "function_block_screen.listen_screen",
-    "unique": false
-}
