--- GTMBplugin_B/Presets/nbteditor.preset
@@ -1,34 +1,0 @@
-{
-    "autoCreate": false,
-    "bindParentAutoScale": false,
-    "bindParentOffset": [
-        0,
-        0,
-        0
-    ],
-    "boxDataList": [],
-    "childParts": [],
-    "childPresets": [],
-    "classType": "UIPreset",
-    "createUIMethod": "PushScreen",
-    "dimension": 0,
-    "forceLoad": true,
-    "isBindParent": false,
-    "isHud": true,
-    "mNode": null,
-    "name": "nbteditor",
-    "notifyRemove": true,
-    "objId": "c0192ad451ad49aebf0b8d85f42c64d0",
-    "preLoad": true,
-    "presetId": "500d40a5-0199-44fc-8fc7-a8a11e68a4c0",
-    "presetType": "界面预设",
-    "previewEnable": true,
-    "showInEditor": true,
-    "singleton": false,
-    "uiName": "nbteditor",
-    "uiNamespace": "ycgnw",
-    "uiNodeModule": "nbteditor",
-    "uiNodeModulePath": "gtmbPlugin\\uiScript\\nbteditor",
-    "uiNodeScreen": "nbteditor.main_closable",
-    "unique": false
-}
