--- GTMBplugin_B/Presets/struimport.preset
@@ -1,34 +1,0 @@
-{
-    "autoCreate": false,
-    "bindParentAutoScale": false,
-    "bindParentOffset": [
-        0,
-        0,
-        0
-    ],
-    "boxDataList": [],
-    "childParts": [],
-    "childPresets": [],
-    "classType": "UIPreset",
-    "createUIMethod": "PushScreen",
-    "dimension": 0,
-    "forceLoad": true,
-    "isBindParent": false,
-    "isHud": true,
-    "mNode": null,
-    "name": "struimport",
-    "notifyRemove": true,
-    "objId": "3c1dda63b67b4ace95857cf9b0a6dcaf",
-    "preLoad": true,
-    "presetId": "87a570a3-d1b4-4321-b3c9-83c1f0ba3925",
-    "presetType": "界面预设",
-    "previewEnable": true,
-    "showInEditor": true,
-    "singleton": false,
-    "uiName": "struimport",
-    "uiNamespace": "ycgnw",
-    "uiNodeModule": "importstrulogic",
-    "uiNodeModulePath": "gtmbPlugin\\uiScript\\importstrulogic",
-    "uiNodeScreen": "structureimport.main_closable",
-    "unique": false
-}
