--- GTMBplugin_B/gtmbPlugin/Parts/FunctionBlock/FunctionBlock.part
@@ -1,11 +1,0 @@
-{
-    "bpType": "FunctionBlockPart",
-    "changes": {},
-    "classType": "PartInstance",
-    "clientPartType": null,
-    "customVariables": [],
-    "directoryPath": "FunctionBlock",
-    "metaType": "FunctionBlockPartMeta",
-    "objId": "d99a70c92bad47189290cf8da14a6783",
-    "partType": "FunctionBlockPart"
-}
