+++ GTMBplugin_B/gtmbPlugin/uiScript/listenBlockScreen.py
@@ -1,22 +1,21 @@
 # -*- coding: utf-8 -*-
 import mod.client.extraClientApi as clientApi
-from gtmbPlugin.function_nodes import listeners
+from function_nodes import listeners
 ViewBinder = clientApi.GetViewBinderCls()
 ViewRequest = clientApi.GetViewViewRequestCls()
 ScreenNode = clientApi.GetScreenNodeCls()
 CFClient = clientApi.GetEngineCompFactory()
-clientSystem = clientApi.GetSystem('Minecraft', 'preset')
+clientSystem = clientApi.GetSystem('gtmbPlugin', 'functionBlockClientSystem')
 
 class listenBlockScreen(ScreenNode):
 	def __init__(self, namespace, name, param):
 		ScreenNode.__init__(self, namespace, name, param)
+		self.currentMode = param['mode']
 
 	def Create(self):
 		"""
 		@description UI创建成功时调用
 		"""
-		clientSystem.NotifyToServer('getNodeBlock', {})
-		clientSystem.ListenForEvent('Minecraft', 'preset', 'nodeBlockCallBack', self, self.gotNodeBlock)
 
 		get_button = lambda path: self.GetBaseUIControl(path).asButton()
 		exist_buttons = ['/background/close', 'background/close_dont_save']
@@ -25,21 +24,17 @@ class listenBlockScreen(ScreenNode):
 		get_button('/background/close').SetButtonTouchUpCallback(self.save)
 		get_button('background/close_dont_save').SetButtonTouchUpCallback(self.close)
 		comboBox = self.GetBaseUIControl('/leftside_panel/node_selector').asNeteaseComboBox()
-		for i in listeners.keys():
-			comboBox.AddOption(i, None, {'tip': listeners.get(i, '无对应内容')})
+		for i in listeners:
+			comboBox.AddOption(i, None, {'tip': listeners[i]})
 		comboBox.RegisterSelectItemCallback(self.load_current_node)
+		self.load_current_node(None, 
+						 self.currentMode or '选择一个模式', 
+						 {'tip': listeners.get(self.currentMode, 'None') if self.currentMode else '选择一个模式并查看其说明...'})
 
 	def save(self, args):
 		selected = self.GetBaseUIControl('/leftside_panel/node_selector/comboBoxTitlePanel/titleContentPanel/titleLabel').asLabel().GetText()
 		clientSystem.NotifyToServer('NodeBlockSave', {'nodeType': selected})
 		self.close(None)
-
-	def gotNodeBlock(self, args):
-		if args['type']:
-			self.load_current_node(None, args['type'], {'tip': listeners.get(args['type'], 'None')})
-		else:
-			self.load_current_node(None, '选择一个模式', {'tip': '选择一个模式并查看其说明...'})
-		clientSystem.UnListenForEvent('Minecraft', 'preset', 'nodeBlockCallBack', self, self.gotNodeBlock)
 
 	def load_current_node(self, index, showName, userData):
 		self.GetBaseUIControl('/node_tip').asLabel().SetText(userData['tip'])
